test: Rename testsuite/clean.rs to testsuite/clean_legacy_layout.rs
