+++ tests/testsuite/clean_legacy_layout.rs
