+++ tests/testsuite/main.rs
@@ -73,7 +73,7 @@ mod cargo_yank;
 mod cfg;
 mod check;
 mod check_cfg;
-mod clean;
+mod clean_legacy_layout;
 mod clean_new_layout;
 mod collisions;
 mod compile_time_deps;
